+++ packages/core/src/sessions/thinkingActivities.ts
@@ -1,0 +1,103 @@
+export const THINKING_ACTIVITIES = [
+  "Booping",
+  "Crunching",
+  "Digging",
+  "Fetching",
+  "Inferring",
+  "Indexing",
+  "Juggling",
+  "Noodling",
+  "Peeking",
+  "Percolating",
+  "Poking",
+  "Pondering",
+  "Scanning",
+  "Scrambling",
+  "Sifting",
+  "Sniffing",
+  "Spelunking",
+  "Tinkering",
+  "Unraveling",
+  "Decoding",
+  "Trekking",
+  "Sorting",
+  "Trimming",
+  "Mulling",
+  "Surfacing",
+  "Rummaging",
+  "Scouting",
+  "Scouring",
+  "Threading",
+  "Hunting",
+  "Swizzling",
+  "Grokking",
+  "Hedging",
+  "Scheming",
+  "Unfurling",
+  "Puzzling",
+  "Dissecting",
+  "Stacking",
+  "Snuffling",
+  "Hashing",
+  "Clustering",
+  "Teasing",
+  "Cranking",
+  "Merging",
+  "Snooping",
+  "Rewiring",
+  "Bundling",
+  "Linking",
+  "Mapping",
+  "Tickling",
+  "Flicking",
+  "Hopping",
+  "Rolling",
+  "Zipping",
+  "Twisting",
+  "Blooming",
+  "Sparking",
+  "Nesting",
+  "Looping",
+  "Wiring",
+  "Snipping",
+  "Zoning",
+  "Tracing",
+  "Warping",
+  "Twinkling",
+  "Flipping",
+  "Priming",
+  "Snagging",
+  "Scuttling",
+  "Framing",
+  "Sharpening",
+  "Flibbertigibbeting",
+  "Kerfuffling",
+  "Dithering",
+  "Discombobulating",
+  "Rambling",
+  "Befuddling",
+  "Waffling",
+  "Muckling",
+  "Hobnobbing",
+  "Galumphing",
+  "Puttering",
+  "Whiffling",
+  "Thinking",
+] as const;
+
+export function pickThinkingActivity(randomValue: number): string {
+  const bounded = Math.max(0, Math.min(randomValue, 0.999999999999));
+  return THINKING_ACTIVITIES[Math.floor(bounded * THINKING_ACTIVITIES.length)];
+}
+
+export function pickNextThinkingActivity(
+  current: string,
+  randomValue: number,
+): string {
+  const picked = pickThinkingActivity(randomValue);
+  if (picked !== current || THINKING_ACTIVITIES.length <= 1) return picked;
+  const index = THINKING_ACTIVITIES.indexOf(
+    current as (typeof THINKING_ACTIVITIES)[number],
+  );
+  return THINKING_ACTIVITIES[(index + 1) % THINKING_ACTIVITIES.length];
+}
