+++ packages/ui/src/features/posthog-mcp/utils/posthog-exec-display.test.ts
@@ -1,9 +1,9 @@
-import { describe, expect, it } from "vitest";
 import {
   formatPosthogExecBody,
   getPostHogExecDisplay,
   isPostHogExecTool,
-} from "./posthog-exec-display";
+} from "@posthog/core/sessions/posthogExecDisplay";
+import { describe, expect, it } from "vitest";
 
 describe("isPostHogExecTool", () => {
   it("matches the bare posthog exec tool", () => {
